dead-code: bin links the lib instead of recompiling it (Issue #475)
src/main.rs declared its own mod tree, so the rust_scorer bin target
compiled a second, independent copy of every module. Anything used only
from the lib side was dead in the bin copy, which forced 17
#[allow(dead_code)] suppressions and kept a dozen pub markers pinned.

CLI logic moved to src/cli.rs (a module of the lib); main.rs is now a
thin shim over rust_scorer::cli::main. All 17 suppressions are removed
and ten pub items that existed only to dodge dead_code are pub(crate),
so the lint is armed crate-wide. No behaviour or CLI-contract change.

ScoringPath, RecordSampler and parse_sample_rate stay pub, documented
inline: the first has a test-only variant, the other two are consumed by
doctests (external crates).

Vibe-Coder-Run-Id: vibe-ms5ya3z5-6e9fa2
